+++ CHANGELOG.rst
@@ -9,6 +9,8 @@ Changelog
 
 version 0.12.0-dev
 ------------------
++ Properly name percentiles as such in the sequence length distribution rather
+  than using N50 nomenclature which is not correct.
 + Fix a bug where BAM files with missing quality sequences were inproperly 
   handled.
 + Update internal UniVec database to version from November 21st 2023.